Create Day 6.cpp
Making program for Day 5 robust, such that when a user inputs a wrong choice, he or she is prompted to enter a correct one until they get it right. Also making a modification to the program such that the user can choose between computing sum/product up till desired number with or without the multiples of 3 and 5. Happy coding!🎉