Create Day_17.cpp
You have been asked to product a "Monthly Sales Report" for your neighbor's retail chain of appliance stores. First, create a data file infile.txt containing sales data for his 7 stores. (Add 5 more data values for a total 7 stores each with 5 items each)     
The first data item is the store name. It should be store in an array of 7 strings.
The next 5 float values are the sales totals for each store over the the first five months of the year. These values should be stored in a two-dimensional array (7 by 5). 
In the Get_store_data function, you must use a nested loop to read and store the data into the array of store names and into the two-dimensional store sales array.
In the Compute_sales_total function, the store index is used to read the sales figures in the indicated row of the sales array. Using a for statement, this function should total the sales amounts and return this total to main.
Your main function should first call the Get_store_data function. Then, using a for loop, the main function should call the Compute_sales_total function sending it each of the seven indices for the rows of the sales array. Of course, the sales totals for each of the stores must be stored in main in a store totals array.
Finally, the main function must call the Display_Store_Data function send it the store sales array, the store names array, and the store totals array. This function simply prints a heading and then, using a counter-controlled loop, print the corresponding data from each array to produce the output